#5 feat: added auth feature

+++ src/users/users.controller.ts
@@ -8,40 +8,90 @@ import {
   Post,
   ParseIntPipe,
   ValidationPipe,
+  UseGuards,
+  Request,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common'
 import { UsersService } from './users.service'
 import { CreateUserDto } from './dto/create-user.dto'
 import { UpdateUserDto } from './dto/update-user.dto'
+import { JwtAuthGuard } from 'src/guards/jwt-auth.guard'
 
-@Controller('users') // /users
+@Controller('users')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
-  @Get('all_users') // /users/all_users
-  findAll() {
+  @UseGuards(JwtAuthGuard)
+  @Get('all_users')
+  findAll(@Request() req) {
+    if (req.user.user_role != 'ADMIN') {
+      throw new HttpException(
+        'Invalid authentication level',
+        HttpStatus.FORBIDDEN,
+      )
+    }
     return this.usersService.findAll()
   }
 
-  @Get('user/:id') // /users/user/:id
-  findOne(@Param('id', ParseIntPipe) id: number) {
-    return this.usersService.findOne(id)
+  @UseGuards(JwtAuthGuard)
+  @Get('user/:id')
+  findOne(@Request() req, @Param('id', ParseIntPipe) id: number) {
+    if (req.user.user_role === 'CONTRIBUTOR' && req.user.user_id === id) {
+      return this.usersService.findOne(id)
+    } else if (req.user.user_role === 'CONTRIBUTOR' && req.user.user_id != id) {
+      throw new HttpException(
+        'Invalid authentication level',
+        HttpStatus.FORBIDDEN,
+      )
+    } else if (req.user.user_role === 'ADMIN') {
+      return this.usersService.findOne(id)
+    }
   }
 
-  @Post('create') // /users/create
-  create(@Body(ValidationPipe) createUserDto: CreateUserDto) {
+  @UseGuards(JwtAuthGuard)
+  @Post('create')
+  create(@Request() req, @Body(ValidationPipe) createUserDto: CreateUserDto) {
+    if (req.user.user_role != 'ADMIN') {
+      throw new HttpException(
+        'Invalid authentication level',
+        HttpStatus.FORBIDDEN,
+      )
+    }
     return this.usersService.create(createUserDto)
   }
 
-  @Patch('update/:id') // /users/update/:id
+  @UseGuards(JwtAuthGuard)
+  @Patch('update/:id')
   update(
     @Param('id', ParseIntPipe) id: number,
     @Body(ValidationPipe) updateUserDto: UpdateUserDto,
+    @Request() req,
   ) {
-    return this.usersService.update(id, updateUserDto)
+    if (req.user.user_role === 'CONTRIBUTOR' && req.user.user_id != id) {
+      throw new HttpException(
+        'Invalid authentication level',
+        HttpStatus.FORBIDDEN,
+      )
+    } else if (
+      req.user.user_role === 'CONTRIBUTOR' &&
+      req.user.user_id === id
+    ) {
+      return this.usersService.update(id, updateUserDto)
+    } else if (req.user.user_role === 'ADMIN') {
+      return this.usersService.update(id, updateUserDto)
+    }
   }
 
-  @Delete('delete/:id') // /users/delete/:id
-  delete(@Param('id', ParseIntPipe) id: number) {
+  @UseGuards(JwtAuthGuard)
+  @Delete('delete/:id')
+  delete(@Request() req, @Param('id', ParseIntPipe) id: number) {
+    if (req.user.user_role != 'ADMIN') {
+      throw new HttpException(
+        'Invalid authentication level',
+        HttpStatus.FORBIDDEN,
+      )
+    }
     return this.usersService.delete(id)
   }
 }
